requirement.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
boto3==1.9.179
botocore==1.12.179
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector-python==8.0.18
mysqlclient==1.4.4
numpy==1.17.4
pandas==0.25.3
protobuf==3.10.0
psycopg2==2.8.3
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyyml==0.0.2
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.3
Werkzeug==0.16.0
xmltodict==0.12.0
